[chore][pkg/stanza] Remove defunct splittest framework
