+++ pkg/stanza/split/splittest/splittest.go
@@ -5,107 +5,118 @@ package splittest // import "github.com/open-telemetry/opentelemetry-collector-c
 
 import (
 	"bufio"
-	"io"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
 )
 
-// state is going to keep processing state of the testReader
-type state struct {
-	ReadFrom  int
-	Processed int
+type Step struct {
+	waitAtEOF func() bool
+	validate  func(t *testing.T, advance int, token []byte, err error)
 }
 
-// testReader is a testReader which keeps state of readed and processed data
-type testReader struct {
-	State *state
-	Data  []byte
-}
+var noWait = func() bool { return false }
 
-// newTestReader creates testReader with empty state
-func newTestReader(data []byte) testReader {
-	return testReader{
-		State: &state{
-			ReadFrom:  0,
-			Processed: 0,
+func ExpectReadMore() Step {
+	return Step{
+		waitAtEOF: noWait,
+		validate: func(t *testing.T, advance int, token []byte, err error) {
+			assert.True(t, needMoreData(advance, token, err))
 		},
-		Data: data,
 	}
 }
 
-// Read reads data from testReader and remebers where reading has been finished
-func (r testReader) Read(p []byte) (n int, err error) {
-	// return eof if data has been fully readed
-	if len(r.Data)-r.State.ReadFrom == 0 {
-		return 0, io.EOF
-	}
-
-	// iterate over data char by char and write into p
-	// until p is full or no more data left to read
-	i := 0
-	for ; i < len(r.Data)-r.State.ReadFrom; i++ {
-		if i == len(p) {
-			break
-		}
-		p[i] = r.Data[r.State.ReadFrom+i]
-	}
-
-	// update state
-	r.State.ReadFrom += i
-	return i, nil
+func ExpectToken(expectToken string) Step {
+	return ExpectAdvanceToken(len(expectToken), expectToken)
 }
 
-// Reset resets testReader state (sets last readed position to last processed position)
-func (r *testReader) Reset() {
-	r.State.ReadFrom = r.State.Processed
-}
-
-func (r *testReader) splitFunc(split bufio.SplitFunc) bufio.SplitFunc {
-	return func(data []byte, atEOF bool) (advance int, token []byte, err error) {
-		advance, token, err = split(data, atEOF)
-		r.State.Processed += advance
-		return
+func ExpectAdvanceToken(expectAdvance int, expectToken string) Step {
+	return Step{
+		waitAtEOF: noWait,
+		validate: func(t *testing.T, advance int, token []byte, err error) {
+			assert.Equal(t, expectAdvance, advance)
+			assert.Equal(t, []byte(expectToken), token)
+			assert.NoError(t, err)
+		},
 	}
 }
 
-type TestCase struct {
-	Name                 string
-	Pattern              string
-	OmitPattern          bool
-	Input                []byte
-	ExpectedTokens       []string
-	ExpectedError        error
-	Sleep                time.Duration
-	AdditionalIterations int
+func ExpectAdvanceNil(expectAdvance int) Step {
+	return Step{
+		waitAtEOF: noWait,
+		validate: func(t *testing.T, advance int, token []byte, err error) {
+			assert.Equal(t, expectAdvance, advance)
+			assert.Equal(t, []byte(nil), token)
+			assert.NoError(t, err)
+		},
+	}
 }
 
-func (tc TestCase) Run(splitFunc bufio.SplitFunc) func(t *testing.T) {
-	reader := newTestReader(tc.Input)
+func ExpectError(expectErr string) Step {
+	return Step{
+		waitAtEOF: noWait,
+		validate: func(t *testing.T, advance int, token []byte, err error) {
+			assert.EqualError(t, err, expectErr)
+		},
+	}
+}
 
+func Eventually(step Step, maxTime time.Duration, tick time.Duration) Step {
+	var waited time.Duration
+	step.waitAtEOF = func() bool {
+		time.Sleep(tick)
+		waited += tick
+		return waited < maxTime
+	}
+	return step
+}
+
+func New(splitFunc bufio.SplitFunc, input []byte, steps ...Step) func(*testing.T) {
 	return func(t *testing.T) {
-		var tokens []string
-		for i := 0; i < 1+tc.AdditionalIterations; i++ {
-			// sleep before next iterations
-			if i > 0 {
-				time.Sleep(tc.Sleep)
-			}
-			reader.Reset()
-			scanner := bufio.NewScanner(reader)
-			scanner.Split(reader.splitFunc(splitFunc))
-			for {
-				ok := scanner.Scan()
-				if !ok {
-					assert.Equal(t, tc.ExpectedError, scanner.Err())
-					break
-				}
-				tokens = append(tokens, scanner.Text())
-			}
-		}
+		var offset int
+		for _, step := range append(steps, ExpectReadMore()) {
+			// Split funcs do not have control over the size of the
+			// buffer so must be able to ask for more data as needed.
+			// Start with a tiny buffer and grow it slowly to ensure
+			// the split func is capable of asking appropriately.
+			var bufferSize int
+			var atEOF bool
+			var advance int
+			var token []byte
+			var err error
 
-		assert.Equal(t, tc.ExpectedTokens, tokens)
+			for needMoreData(advance, token, err) && (!atEOF || step.waitAtEOF()) {
+				// Grow the buffer at a slow pace to ensure that we're
+				// exercising the split func's ability to ask for more data.
+				bufferSize = 1 + bufferSize + bufferSize/8
+				data := make([]byte, 0, bufferSize)
+				if offset+bufferSize >= len(input) {
+					atEOF = true
+					data = append(data, input[offset:]...)
+				} else {
+					data = append(data, input[offset:offset+bufferSize]...)
+				}
+				advance, token, err = splitFunc(data, atEOF)
+				// t.Errorf("\nbuffer: %d, advance: %d, token: %q, err: %v", bufferSize, advance, token, err)
+			}
+			offset += advance
+			step.validate(t, advance, token, err)
+		}
 	}
+}
+
+func needMoreData(advance int, token []byte, err error) bool {
+	return advance == 0 && token == nil && err == nil
+}
+
+// ScanLinesStrict behaves like bufio.ScanLines except EOF is not considered a line ending.
+func ScanLinesStrict(data []byte, atEOF bool) (advance int, token []byte, err error) {
+	advance, token, err = bufio.ScanLines(data, atEOF)
+	if advance == len(token) {
+		return 0, nil, nil
+	}
+	return
 }
 
 func GenerateBytes(length int) []byte {
